Fix 45 failing tests on Windows.
On Windows, os.Getwd() returns '\' style paths, but the
filenames returned from function.FileLine(pc) return
'/' style paths. So the check to see if the long-path
starts with Getwd fails.
Add an init() function so that we can
a) add the trailing '/', rather than doing it on every call
   to nicePath()
b) Switches '\' delimiters to '/' so that they actually match.