Support Go Modules
Unlike previous attempts at adding a go.mod file (see #108), this keeps
the existing gopkg.in based module name so that everything continues to
work with older versions of Go, or with projects that are using the
existing import. In future, a semver compatible tag can be added (eg.
v0.0.1 or v1.0.0) which will allow this package to be pinned to by
projects using Go Modules in a more readable way.